Capacity Miss 2/2
Kinda funzionante, mancano piu policy e bisogna fare più test(tipo con file lockati)

+++ include/SERVER/filesystem.h
@@ -41,9 +41,9 @@ void fsDestroy(FileSystem *fs);
 int openFile(const char *pathname, int flags, FileDescriptor **fd, FileSystem *fs);
 int closeFile(FileDescriptor *fd, FileSystem *fs);
 
-int readFile(FileDescriptor *fd, void **buf, size_t size, FileSystem *fs);
-int writeFile(FileDescriptor *fd, void *buf, size_t size, FileSystem *fs);
-int appendToFile(FileDescriptor *fd, void *buf, size_t size, FileSystem *fs);
+int readFile(FileDescriptor *fd, void **buf, uint64_t size, FileSystem *fs);
+int writeFile(FileDescriptor *fd, void *buf, uint64_t size, FileSystem *fs);
+int appendToFile(FileDescriptor *fd, void *buf, uint64_t size, FileSystem *fs);
 
 int readNFiles(int N, FileContainer **buf, FileSystem *fs);
